Updated lists to pull from avaialabe games and agents
